feat: move carousel to own component
